move TypedLinkException to ferry_exec
