+++ ferry_exec/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3]
+
+- add `TypedLinkException`
+
 ## [0.0.2]
 
 - update dependencies
