+++ ferry_exec/lib/ferry_exec.dart
@@ -3,3 +3,4 @@ export 'package:ferry_exec/src/fragment_request.dart';
 export 'package:ferry_exec/src/operation_request.dart';
 export 'package:ferry_exec/src/operation_response.dart';
 export 'package:ferry_exec/src/typed_link.dart';
+export 'package:ferry_exec/src/exceptions.dart';
