+++ ferry_exec/lib/src/exceptions.dart
@@ -1,0 +1,16 @@
+import 'package:ferry_exec/ferry_exec.dart';
+import 'package:gql_link/gql_link.dart';
+
+/// A base class for exceptions thrown by [TypedLink]s
+class TypedLinkException extends LinkException {
+  const TypedLinkException(
+    dynamic originalException,
+  ) : super(originalException);
+
+  @override
+  bool operator ==(Object o) =>
+      o is TypedLinkException && o.originalException == originalException;
+
+  @override
+  int get hashCode => originalException.hashCode;
+}
