+++ ferry_exec/pubspec.yaml
@@ -1,5 +1,5 @@
 name: ferry_exec
-version: 0.0.2
+version: 0.0.3
 homepage: https://ferrygraphql.com/
 description: A strongly typed execution interface for GraphQL operations
 repository: https://github.com/gql-dart/ferry
